requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
einops==0.8.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.2.0
ftfy==6.2.0
humanfriendly==10.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.3
Jinja2==3.1.3
jmespath==0.10.0
kiwisolver==1.4.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mat4py==0.6.0
matplotlib==3.7.5
mdurl==0.1.2
mmcv==2.1.0
-e git+https://github.com/open-mmlab/mmdetection.git@cfd5d3a985b0249de009b67d04f37263e11cdf3d#egg=mmdet
mmengine==0.10.4
mmpretrain==1.2.0
-e git+https://github.com/open-mmlab/mmsegmentation.git@b040e147adfa027bbc071b624bedf0ae84dfc922#egg=mmsegmentation
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.0.3
pillow==10.2.0
platformdirs==4.2.2
prettytable==3.11.0
protobuf==5.28.2
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2023.4
pywin32==306
PyYAML==6.0.2
regex==2024.7.24
requests==2.28.2
rich==13.4.2
scipy==1.10.1
shapely==2.0.5
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
tomli==2.0.1
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.65.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.2.13
yapf==0.40.2
zipp==3.20.0